fhir-interoperability/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
certifi==2022.12.7
    # via requests
charset-normalizer==3.1.0
    # via requests
dnspython==2.3.0
    # via email-validator
email-validator==1.3.1
    # via pydantic
fhir-resources==6.5.0
    # via -r requirements.in
idna==3.4
    # via
    #   email-validator
    #   requests
pydantic[email]==1.10.7
    # via fhir-resources
requests==2.28.2
    # via -r requirements.in
typing-extensions==4.5.0
    # via pydantic
urllib3==1.26.15
    # via requests
